requirements.txt
flask==3.0.0
langchain-google-genai==1.0.10
langchain>=0.2.0
langchain-core>=0.2.33,<0.3
python-dotenv==1.0.0
google-generativeai>=0.7.0,<0.8.0
